complete provision instructions
cause this is not enough to use this module from the public registry

```
module "sql-db" {
  source  = "GoogleCloudPlatform/sql-db/google"
  version = "3.1.0"
}
```

This warn help us, but can be more explicit :
This module version (...) has no root configuration. A module with no root configuration cannot be used directly.
Use the submodules dropdown above to view the 4 submodules defined within this module.

So add a small usage guideline in the readme with remote usage can help us