Add Suffix to backup configurations
The current setup sets an unique name for sql_instance for the backup
jobs. However, it is possible that an user might want more than one
backup job - for instance, a weekly, a monthly and a yearly.

The current setup does not allow having diff jobs with different
retentions otherwise.