Switch from `lookup()` to `coalesce()` function.
The `lookup()` func was causing errors such as
`Invalid value for "str" parameter: argument must not be null.` on the
`region` and `encryption_key_name` fields.

I believe this is related to
https://github.com/hashicorp/terraform/issues/32157.

By switching to `coalesce()`, the first non-null value will be returned,
which I believe is the intended behaviour with the previous `lookup()`
usage.