feat: add support for server_ca_mode and custom SANs (#714)
Add support for configuring server_ca_mode and server_ca_pool in the
PostgreSQL module's IP configuration. This enables custom certificate
authority management for CloudSQL instances.

Also add support for custom Subject Alternative Names (SANs) to allow
custom DNS names to be included in the server certificate.

This fixes an issue with cloud-sql-proxy version after 2.15.2 that fails
to verify certificates when using custom domain names, as described in:
https://github.com/GoogleCloudPlatform/cloud-sql-proxy/issues/2425

Increase minimum Google provider version to 6.31.0, as the
custom_subject_alternative_names field was introduced in this version:
https://github.com/hashicorp/terraform-provider-google/releases/tag/v6.31.0